kakao/naver login& change Password& edit profile

+++ controller/videoController.js
@@ -40,8 +40,11 @@ export const postUpload = async (req, res) => {
   const newVideo = await videoModel.create({
     fileUrl: path,
     title,
-    description
+    description,
+    creator: req.user.id
   });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
   res.redirect(routes.videoDetail(newVideo.id));
 };
 export const videos = (req, res) => {
@@ -52,7 +55,8 @@ export const videoDetail = async (req, res) => {
     params: { id }
   } = req;
   try {
-    const video = await videoModel.findById(id);
+    const video = await videoModel.findById(id).populate("creator");
+    console.log(video);
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
     res.redirect(routes.home);
@@ -62,7 +66,11 @@ export const getEditVideo = async (req, res) => {
   const { id } = req.params;
   try {
     const video = await videoModel.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    if (video.creator !== req.user._id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
     res.redirect(routes.videoDetail(id));
   }
@@ -84,8 +92,13 @@ export const deleteVideo = async (req, res) => {
     params: { id }
   } = req;
   try {
-    await videoModel.findOneAndDelete({ _id: id });
-    res.redirect(routes.home);
+    const video = await videoModel.findById(id);
+    if (video.creator !== req.user._id) {
+      throw Error();
+    } else {
+      await videoModel.findOneAndDelete({ _id: id });
+      res.redirect(routes.home);
+    }
   } catch (error) {
     res.redirect(routes.videoDetail(id));
   }
